refactor: make it easier to add new diagnostics
